added possibility to filter nodes
* There is alread the "nodeCanvasObject" callback which allows to
paint custom canvas objects as nodes. Unfortunately,
this callback gets called for _all_ nodes.

This extension adds a new option to filter nodes on a provided
property name, so that the "nodeCanvasObject" callback gets only
called for specific nodes. For the remaining nodes, the default
node object will be used.